# BackendNode/Backend/app/configs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "P0136. GIS - Prototype Development",
    "description": "- Use opensource\n - Using AWS\n - First demo Aug - 01\n - Release POC end of September\n - Start date: 07/11/2019\n - End date: 09/20/2019"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status of role that we want to find(active or inactive) ex: status=active.",
            "type": "string"
          },
          {
            "name": "sort_field",
            "in": "query",
            "description": "Field of role that we want to sort ex: sort_field=status.",
            "type": "string"
          },
          {
            "name": "sort_type",
            "in": "query",
            "description": "Type of Field that we want to sort(asc or desc) ex: sort_type=asc.",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number, starting with 1. ex: page=1.",
            "type": "number"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Requested response Page size. Minimum, maximum and default is defined by Provider. Minimum can never be lower than 1 ex: pageSize=10.",
            "type": "number"
          },
          {
            "name": "searchValue",
            "in": "query",
            "description": "Search role by name include searchValue.",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "summary": "Get all roles in system",
        "responses": {
          "200": {
            "description": "Succesfully retrieved resource.",
            "schema": {
              "$ref": "#/definitions/RoleResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "post": {
        "summary": "Create new role in system",
        "tags": ["Roles"],
        "parameters": [
          {
            "name": "role",
            "in": "body",
            "description": "Role that we want to create",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New role is created",
            "schema": {
              "$ref": "#/definitions/RoleResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/roles/{roleId}": {
      "parameters": [
        {
          "name": "roleId",
          "in": "path",
          "required": true,
          "description": "ID of role that we want to find",
          "type": "string"
        },
        {
          "name": "Authorization",
          "in": "header",
          "description": "An authorization header .",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "tags": ["Roles"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "summary": "Get role with given ID",
        "responses": {
          "200": {
            "description": "Role is found",
            "schema": {
              "$ref": "#/definitions/RoleResponse"
            }
          },
          "400": {
            "description": "Role is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete role with given ID",
        "tags": ["Roles"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Role is deleted",
            "schema": {
              "$ref": "#/definitions/RoleResponse"
            }
          },
          "400": {
            "description": "Role is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "put": {
        "summary": "Update role with give ID",
        "tags": ["Roles"],
        "parameters": [
          {
            "name": "role",
            "in": "body",
            "description": "Role with new values of properties",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Role is updated",
            "schema": {
              "$ref": "#/definitions/RoleResponse"
            }
          },
          "400": {
            "description": "Role is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "summary": "Login into the system with userName & password",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Status that we want to change",
            "schema": {
              "properties": {
                "userName": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticate success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "access_token": {
                        "type": "string"
                      },
                      "refresh_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/auth/refresh_token": {
      "post": {
        "summary": "Get new token",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "body",
            "schema": {
              "properties": {
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticate success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "access_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout the system",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "body",
            "schema": {
              "properties": {
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": []
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Role": {
      "required": ["id", "name"],
      "properties": {
        "id": {
          "type": "number",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": ["active", "inactive"]
        }
      }
    },
    "RoleResponse": {
      "properties": {
        "status": {
          "$ref": "#/definitions/Status"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          }
        }
      }
    },
...
    "Status": {
      "properties": {
        "error": {
          "type": "boolean",
          "example": false
        },
        "code": {
          "type": "number"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "NotFound": {
      "properties": {
        "status": {
          "$ref": "#/definitions/Status"
        },
        "results": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": []
        }
      }
    }
  }
}